requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
Django==5.0.4
django-cors-headers==4.3.1
django-googlesearch==0.2
djangorestframework==3.15.1
dnspython==2.6.1
docutils==0.21.2
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
langcodes==3.4.0
language_data==1.2.0
lockfile==0.12.2
lxml==5.2.1
marisa-trie==1.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pid==3.0.4
preshed==3.0.9
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
PySocks==1.7.1
python-daemon==3.0.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.20.0
service==0.6.0
setproctitle==1.3.3
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
